refactor(kv-router): move bench_utils from src/ into benches/common/
bench_utils was gated behind #[cfg(feature = "bench")] in the library crate
but only used by bench harnesses. Move all types (SequenceData,
generate_sequences, NoopSequencePublisher, SimpleWorkerConfig, median) into
benches/common/mod.rs where they naturally belong, and delete the src module.
